Add a more compact representation of usage stats.
We are replaceing the package name in the activity name with a * iff the activity is in the
same package, otherwise the activity name is pritned out in full.
This small change will remove a lot of bytes (in the order of kilobytes for a real log) from the logged data on the network and downstream processing,
since the package name is repeated in almost all cases.

 An exampe of the new format is here:
 DUMP OF SERVICE usagestats:
 D:4,20090813
 P:com.android.launcher,4,155456
 A:*.Launcher,4,0,0,0,0,0,0,0,0,0,2
 P:com.android.browser,1,6724
 A:*.BrowserActivity,1,0,0,0,0,0,0,0,0,0,0
 A:*.CombinedBookmarkHistoryActivity,1,0,0,0,0,0,0,0,0,0,1
 P:com.google.android.apps.maps,1,2219
 A:com.google.android.maps.MapsActivity,1,0,0,0,0,0,0,0,0,0,0
 P:com.android.contacts,1,0
 A:*.DialtactsActivity,1,0,0,0,0,0,0,0,0,0,1