Implement new title bar hiding/showing behavior.
Remove the code that hides the title bar when the page finishes loading.
When the user double taps, if it did not result in changing the viewing
mode (i.e. on a mobile site), send a message to the client so we can
toggle the title bar's visibility.  Required for a change to
packages/apps/Browser

Change-Id: I0040fd183ce5673a4069e943f38eb31599c0ca7d