Fix bug that would leave old windows around.
The problem was instigated by another fix I made to keep
windows around while their tokens are animating or preparing to
animate.  This then hit an issue where if the window's process
died while its token was getting ready to animate, we would
immediately hide the token, and when we then went to start
the token animation we wouldn't do anything, leaving the old
dummy animation in place and not allowing it to run.

Now we clear the animation before hiding the token, to ensure
the dummy animation is gone at this point.

Change-Id: If83fadcce6815bc545c80fbdb82fe6972bbdf5ef