accelerate consecutive flings
Use the fact that the scroller keeps track of the current velocity
to accelerate subsequent flings. The current velocity is added to
the new velocity if the flings are in the same general direction
and if the fling animation hasn't been aborted.

This makes it easier to move to the top or bottom of very large
pages (either in overview mode or reading mode) by taking
consecutive flings as a cue from the user that they want to
page faster. The user can stop the fling with any action as
before.

fixes http://b/issue?id=2066090