don't fling if there's nowhere to go
Add a test to doFling() to return quickly if the fling has
no effect. For simplicity, only test for horizontal or vertical
flings.

The man purpose is to avoid suspending webkit until the fling
animation is complete when the purpose of the touch drag is
to activate a touch event, like a swipe in superpudu.

fixes http://b/issue?id=2052852