Always interpret (0,0) scrolls as top of title bar.
Move the specialized code in NEW_PICTURE_MSG_ID to
setContentScrollTo so that a scroll to (0,0), whether
a result of a NEW_PICTURE_MSG_ID or from an initial
layout (or otherwise) is treated as scrolling to the
top of the content, including showing the title bar.
Scrolls to (0,1) will trigger code that animates the
title bar off screen.

Change-Id: Ia6db6d4620dea2bfe62407c4394a8ead38501b2f