do not merge: fix copy/paste using trackball on passion
This fix is alrady in mr2. It does not affect sholes --
this is a trackball only bug.

The selection caret and arrow now take the scale factor and
the height of the title bar into account when preparing the
canvas for drawing.

Clicking the trackball at the end of selecting text no longer
sends a click event.

Requires a companion fix in external/webkit
Fixes http://b/issue?id=2187591