Fix bugs around removing and attaching WebViews.
When removing a WebView from its Window, remove the WebTextView.
Remove the old attempt to do the same, which was not being called.

When attaching a WebView to a Window, call onWindowFocusChanged so
its state of drawing the selection ring returns.

Fixes http://b/issue?id=2257787 and http://b/issue?id=2260867